src/auth/cloud_auth_manager.py
from dataclasses import dataclass
from typing import Dict, Any, Optional
from datetime import datetime
import boto3
from google.oauth2 import service_account
from azure.identity import DefaultAzureCredential
from ibm_cloud_sdk_core.authenticators import IAMAuthenticator
import os

@dataclass
class CloudCredentials:
    provider: str
    credentials: Any
    expires_at: Optional[datetime] = None

class UnifiedAuthManager:
    def __init__(self):
        self.credentials: Dict[str, CloudCredentials] = {}
        self.clients: Dict[str, Any] = {}
        self.failure_log: list = []

    def setup_aws_credentials(self):
        """Setup AWS credentials from environment or instance profile"""
        try:
            session = boto3.Session()
            # Test credentials
            sts = session.client('sts')
            identity = sts.get_caller_identity()
            self.credentials['aws'] = CloudCredentials('aws', session)
            return True
        except Exception as e:
            self.log_failure("AWS_AUTH_FAILED", str(e))
            return False

    def setup_gcp_credentials(self):
        """Setup GCP service account or application default credentials"""
        try:
            # Try application default credentials first
            credentials = service_account.Credentials.from_service_account_file(
                os.environ.get('GOOGLE_APPLICATION_CREDENTIALS')
            )
            self.credentials['gcp'] = CloudCredentials('gcp', credentials)
            return True
        except Exception as e:
            self.log_failure("GCP_AUTH_FAILED", str(e))
            return False

    def log_failure(self, failure_type: str, details: str):
        """Log authentication failures for research paper documentation"""
        self.failure_log.append({
            'timestamp': datetime.now(),
            'type': failure_type,
            'details': details,
            'sprint': 1,
            'component': 'authentication'
        })
